+++ crates/medea-macro/src/lib.rs
@@ -200,7 +200,7 @@ pub fn enum_delegate(args: TokenStream, input: TokenStream) -> TokenStream {
 /// #    }
 /// # }
 /// #
-///
+/// #
 /// let mut foo = Foo { bar: 0, baz: 0 };
 ///
 /// let bar = Event::Some { new_bar: 1 }.dispatch_with(&mut foo);
@@ -216,9 +216,71 @@ pub fn enum_delegate(args: TokenStream, input: TokenStream) -> TokenStream {
 /// assert_eq!(foo.baz, 3);
 /// assert_eq!(bar, 3);
 /// ```
+///
+/// # Customize `self` type in handler functions (optional)
+///
+/// By default, all handler functions take `&mut Self`, if this doesn't suit
+/// your case, then you can specify the method receiver manually:
+/// `#[dispatchable(self: Rc<Self>)]`, `#[dispatchable(self: &Self)]`.
+///
+/// You can use any type that is a valid `self` receiver, e.g. `self`, `&self`,
+/// `&mut self`, `self: Box<Self>`, `self: Rc<Self>`, `self: Arc<Self>`, or
+/// `self: Pin<P>` (where P is one of the previous, except `Self`).
+///
+/// ```
+/// # use std::rc::Rc;
+/// use medea_macro::dispatchable;
+///
+/// #[dispatchable(self: Rc<Self>)]
+/// enum Event {
+///     Variant,
+/// }
+///
+/// struct Foo;
+/// impl EventHandler for Foo {
+///    type Output = ();
+///
+///    fn on_variant(self: Rc<Self>) {}
+/// }
+///
+/// let foo = Rc::new(Foo);
+///
+/// Event::Variant.dispatch_with(foo);
+/// ```
+///
+/// # Async handlers (optional)
+///
+/// It's possible to make handler methods `async`. Rust doesn't support `async`
+/// trait methods at the moment, that's why [`async_trait`] is used.
+///
+/// ```
+/// use async_trait::async_trait;
+/// use medea_macro::dispatchable;
+///
+/// #[dispatchable(async_trait(?Send))]
+/// enum Event {
+///     Variant,
+/// }
+///
+/// struct Foo;
+/// #[async_trait(?Send)]
+/// impl EventHandler for Foo {
+///    type Output = ();
+///
+///    async fn on_variant(&mut self) {}
+/// }
+///
+/// let mut foo = Foo;
+///
+/// Event::Variant.dispatch_with(&mut foo);
+/// ```
+///
+/// [`async_trait`]: https://docs.rs/async-trait
 #[proc_macro_attribute]
-pub fn dispatchable(_: TokenStream, input: TokenStream) -> TokenStream {
-    dispatchable::derive(input).unwrap_or_else(|e| e.to_compile_error().into())
+pub fn dispatchable(args: TokenStream, input: TokenStream) -> TokenStream {
+    let enum_item = syn::parse_macro_input!(input as dispatchable::Item);
+    let args = syn::parse_macro_input!(args as dispatchable::Args);
+    dispatchable::expand(enum_item, &args)
 }
 
 decl_derive!([JsCaused, attributes(js)] =>